Refactor: HomePost 컴포넌트화 및 데이터 전달 방법 수정
