+++ src/components/comment/CommentBox.jsx
@@ -1,14 +1,15 @@
 import React from 'react';
+import { useParams } from 'react-router-dom';
 import Comment from './Comment';
 import CommentList from './CommentList';
 
-const CommentBox = () => {
-    const postId = location.pathname.split('/')[3];
-    console.log(postId);
+const CommentBox = ({ postUserName }) => {
+    let { post_id } = useParams();
+    console.log(post_id);
     return (
         <>
-            <CommentList postId={postId} />
-            <Comment postId={postId} />
+            <CommentList postId={post_id} />
+        <Comment postId={post_id} postUserName={postUserName} />
         </>
     );
 };
